feat: Implement JetBrains Markdown for rendering and refine Gemini prompt.
